requirement.txt
fastapi
uvicorn
python-multipart
pydantic
torch
torchvision
transformers
numpy
pandas
SimpleITK
Pillow
joblib
scikit-learn
axios